requirements.txt
fastapi
uvicorn
streamlit
numpy
pandas
torch
PyYAML
pydantic
